Improve headers and remove technical names
We now use Kirigami.ListSectionHeaders for showing category names of the
permissions that follow. This looks better and was designed for this
purpose. Further, the technical names (as per Flatpak Metadata) that
showed up as subtitles in permission's delegates are removed since they
were unnecessary. The technical names are still stored in the
FlatpakPermission class because we will likely need it later.