Load non-default, user-changed simple permissions
So far, permissions were loaded directly from the installed ref's
metadata. These were default permissions that the app has. Changes made
by the user to the permissions were not reflected.

With this commit, the default permissions are still loaded from the
metadata. After that, we check if any of the permissions have been
changed and override the default values with the changed ones. Therefore
for all "simple" permissions, the program now displays the actual
values.