Move setPerm function to permission model
So far it was in the KCMFlatpak class, which made it harder to interact
with each individual permission object and made less sense. This caused
the program to crash if the same permission was edited in a single
session twice, since the value had not been changed in the permission
object but was changed in the overrides file. Now, it works correctly.