Have "add perms" buttons with relevant sec headers
So far, the "Add permissions" button was a global one at the bottom of
the permissions page. In this the user would select the permission
category, and the permission would be added with its default value.

This change gives each section header a seperate "+" button. The user
can select the value that they want to add in the combo box (or type the
value out in case of Environment permissions).