Fix positions of sections and have empty ones
Only those sections were visible which had permissions under the
corresponding category. Further, if the user added a permission using
to a category that has none by default, the section would be appended to
the end of the list, thus the sequence of sections (and even the number)
were not the same across different applications. With the implementation
of seperate "add permission" buttons for each relevant section, sections
with no permission could never have any, because the section would not
appear, hence there would no button to add to it.

This commit does a few things: first, it causes the list section headers
to appear even if there are no permissions under it. This causes each
section to have a "add perms" button. Second, it forces all the sections
to appear in a defined sequence.