Don't remove permission from vector until applied
If a non-simple permission that was not granted by default, is
ungranted, then the permission would disappear. In case of builtin
permission, it would re-appear elsewhere, but in case of user-defined
permission, it would be lost. The user would have to click back on the
same application without applying other changes (thereby losing those
other changes) for it to re-appear.

This change fixes these issues: builtin permissions should not disappear
at all, and user-defined ones should still be in the UI (and the model)
until the user confirms the user applies the changes.