Let user change apps without losing unapplied work
Earlier, when the user made changes and then switched apps without
clicking apply, those changes would be lost when the user returned later
to the same application. This commit saves the user's work in a string
member of the reference object, and loads the user's un-applied changes
from this string when the user returns to the application.

When the user applies the changes, only changes made to the _current_
application get applied, instead of changes to all apps getting applied
at once.