Redesign the KCM into basic/advanced sections
Some permissions (categorized as "Basic") should be understandable to
most users, and are likely to be changed more. Other permissions
(categorized as "Advanced") are less understandable and more
"dangerous".

These permissions are now separated into two different pages. This
provides non-technical users with a less cluttered view of permissions
that are more likely to be of interest to them. Filesystem permissions
are categorized as "basic".

The permissions in the "Basic" section are listed simply. Those in the
advanced section are listed under sections.