Push permissions' page on startup with ref: null
Presently, the permissions' page is not pushed automatically when the
the KCM is started. Instead, it waits for the user to select an
application. This caused the header to get cut off. To fix this and see
a "Permissions" header, we must push the permissions' page as soon as
the KCM starts. However, the ref is set to null, so the permissions
still don't appear for any application until the user clicks one.

Fixes #7.