Replace CheckableListItem with BasicListItem as delegate
The previous commit seemed to have fixed the problem of the component
not registering any changes when the box was clicked. However, it turned
out that clicking anywhere on the delegate other than the box would not
reflect the change on the UI, if the permission clicked was:

1. enabled already
2. the first permission to be clicked after selecting the app

Another problem noticed was that clicking the defaults button would
change the permissions back to defaults on the C++ side, but not reflect
those changes accurately on the UI.

This commit replicates the functionality of CheckableListItem and fixes
these issues. This is a temporary solution and once the solution is
found, we must port it back to CheckableListItem.