Create overrides directory if it doesn't exist
If permissions for any Flatpak app has never been changed in the past,
an "overrides" directory will not exist. This directory must exist prior
to the changing of permissions, since the files holding the present
permissions settings are stored inside it. Therefore this commit
creates the directory if it does not already exist.

Fixes #26.