Create overrides directory if it doesn't exist
If permissions for any of the Flatpak apps has never been changed in the past, an "overrides" directory will not exist. This directory must exist prior to changing the permissions, since the files holding the settings will be placed inside it. Therefore this commit creates the directory if it doesn't already exist.

Further, we now return the overrides path just once, during the startup, instead of (as was the case before this commit) calling the function during the construction of each FlatpakReference instance.

Fixes #26.