Replace i18n by QStringLiteral for permission names and categories
Technical names of permissions and their categories will never be
displayed on the UI. For categories, we use the
FlatpakPermission::categoryHeading() method to return a more
understandable string. We never display permission names.

However these two are used to specify the user's configurations in the
overrides file. Thus these should not be translated.