Implement opening specific application page on start
If a Flatpak application ID is passed in as an argument (--args) to
systemsettings or kcmshell, then this KCM will try to load in requested
application page when it's first initialized.

The plan is to implement scrolling down to and focusing on a specific
permission row(s), so doc comment mentions that in advance.

Unfortunately, due to limitations of KCM API, it is not possible to
request a different page on the fly, without having user to unload KCM
(switch to another one) or fully close System Settings app. Hopefully,
this will be revised in KF6.