UI: Add header with current app icon, name & version to the permissions view
Given the way how ListView works with currentIndex, and easy it is to
get it out of sync with kcm's current app index -- especially when it
comes to filtering model with search -- it would be better to show some
identifying information on the permissions page.  Also good for mobile
form-factor where pages might only be displayed one at a time.