KCM: Re-evaluate state of the standard KCM buttons when unloading permissions model
Apply, Reset & Defaults buttons depend on current index, so the state
should be re-evaluated when index changes. Previously it only relied on
model to relay signals through a bunch of layers to trigger it, but it
won't happen if user hits Enter with an empty search results list --
in that case a placeholder page with a null ref will be pushed, and so
setReference will short-circuit without triggering the required relays.

(cherry picked from commit 2427408f1f63383eb1308a2587bbe8d280952727)

(the change that the following paragraph describes was not cherry-picked
because it is not applicable to Plasma/5.27 branch)

Also initializing state after parsing arguments and potentially setting
m_index makes more sense too, even though QML frontend would pick it up
and load permissions page which would trigger the signal otherwise.