Fix highlight states for permission items
These list items were marked as being checkable, but didn't need to be.
This was probably a remnant of a previous attempt to use
Kirigami.CheckableListItem.

Removing that and then resetting the current index on click brings us to
the desired visual styling of the list items highlighting on hover and
click, but not remaining highlighted after that.


(cherry picked from commit 59df62a87f9f511746daf7c1dc1266ebd64a4d77)