FlatpakPermissionModel: Stop shuffling list of possible values for entries
This was probably done to put currently effective value on top to make
ComboBox select it by default. Instead the proper way to deal with this
situation is to use Component.onCompleted and indexOfValue API.

`if` guard ensures that binding is only established for relevant
delegates, and actual Qt.binding() is needed to support reloading or
resetting to defaults.