Implement filesystems entry (de-)serializer
It tries to stay consistent with flatpak-override command. For example,
it is capable of deduplicating various "home", "~" entries based on
their type rather than actual prefix string literal; and also
serializes them back in a "better compatibility with older Flatpak
versions" mode. Such data structure makes it trivial to check whether
a particular path is already in the list of permissions, as it stores
actual path as a separate string.

List of special directories supported by flatpak is stored as an array
with necessary metadata for each type. This makes it trivial to extend
it, should Flatpak ever add more special filesystems.