FlatpakPermissionModel: Add D-Bus "none" policy to the possible values list
Now it can load and display properly in KCM. Note that "enabled" state
is still to be fixed.

The test is needed because the lists of values are currently assembled
in multiple places independently, so we need to ensure consistency. The
plan is to consolidate all those places in one method and get rid of
ValueList role altogether as it is redundant; but that would also
require some degree of testing.