FlatpakPermissionModel: Expose default enabled and value as model roles
IsDefaultEnabled is going to be used for D-Bus (and later other)
sections to make default-enabled entries not uncheckable, because that
does not make sense.

DefaultValue might come handly to implement default settings highlighter.

It does not matter much that we expose them without usage now, as long
as their names and data() are consistent with FlatpakPermission class.