FlatpakPermissionModel: Sort out enabled state for D-Bus entries
If they are specified in defaults or overrides then it means that they
are enabled. Disabling only makes sense for custom overrides
(which temporarily marks them for deletion), and does not make sense
for default entries (the best we could do in such case is to set them
as "none", but we'd better be explicit about it and not have two ways
to achieve the same thing).

Fixes a bug that disabled entries with previous "none" policies ended
up being duplicated in the config.

(cherry picked from commit 2ea4631d21406b38f5c512d1e5a09530258d0d84)