FlatpakPermissionModel: Use enum for list section headers, move i18n to QML
The point of this change is that now we don't have to "unmap" translated
header text when adding new permissions from sections buttons; and also
this new more granular enum can be reused for Basic/Advanced split, so
the old SectionType enum could be dropped along with all its usages.

Unfortunately, the new type could not be used directly for Q_INVOKABLE
methods and Q_SLOTS, so we still need to fallback to int. Another caveat
is that ListView automatically always converts section data to string.