Nuke possibleValues data member and getter from FlatpakPermission class
It was redundant and wasteful to keep it in-memory for every permission
entry. It can be computed at any time based on entry's section type.

Besides, the model already has this kind of method, it just needed to be
marked as static as well.

(cherry picked from commit 5d299024776da417d254f6a34cf2095dcb2ee740)