QML: Drop all custom properties from delegate, require model object instead
The approach to copy all the model roles to custom properties clearly
was a mess, and tend to only get worse over time. Not only we had to
introduce few new ones, but none of them was marked as readonly either,
and it wouldn't support writing back to a model (through setData)
should we ever consider switching to that.

(cherry picked from commit affbf7a9a1ccaa77072f1abe4d589bb159550eac)