FlatpakPermissionModel: Update "settings changed" state when saving non-default entries
Entries which are not enabled by default don't have any meaningful
default value in upstream metadata, but we reuse the field to represent
currently persisted value in user's overrides file. It saves us from
slightly larger refactoring in which logical expressions of the
SettingHighlighter would need to be rewritten with valueType and
(currently unexposed) OverrideValue role in mind.