FlatpakPermissionModel: Merge defaults from a list of overrides
The model now takes a list of all permissions files (metadata + every
override except the very last one which we operate on), and merges them
one by one into an in-memory KConfig object. The rest is unchanged.

While loading default values, we are not interested in keeping or
processing unknown/unparsable entries in any way, which simplifies code
a bit comparing to loadCurrentValues() method.

BUG: 466997