ui: use ellipses and generic icon name for "Add New"
The HIG recommend that buttons that require further user input end with
an ellipsis. The "Add New" button in the permission list view opens a
dialog, but does not end with an ellipsis. In addition, the requested
icon is bqm-add, related to digikam's BatchQueueManager. This is just a
symlink to the standard list-add right now so there are no negative
user-facing consequences, but it would make more sense to use a
semantically appropriate icon name, in case the bqm icons get changed
later.

This change adds the ellipsis to the button, and requests the list-add
icon instead of bqm-add.