ui: separate active and focused state
The sidebar of the FlatPak kcm mixes two mechanisms: it uses the full
highlighted state that's commonly used to show which entry in the main
content is active and to indicate which item has keyboard focus, This is
rather confusing and doesn't interact well with trying to use keyboard
navigation there.

Instead, properly track the active application also in the UI, and make
both tab navigation and arrow navigation between search field and list
work sensibly (like e.g. in Discover).