ui: fix keyboard navigation on permissions page
The permissions page is currently completely unusable with keyboard
navigation; there is no way to actually get focus there. This is largely
because the sidebar and Permissions page look like sidebar/main content,
but are actually separate pages and not in each others tab chain, plus
the usual keyboard navigation challenges of qml, and the franken-mix of
nested qml in widgets of systemsettings in particular.

This change makes it at least mostly functional. We also add some
convenience by eliminating unnecessary tab stops in the list; the
checkboxes are now available directly from the delegates, and other
delegate controls are only available from that delegate, This makes it
quicker to discover and access the controls not linked to specific
delegates. Known limitations: There is an additional uneccesary tab stop
at the beginning of the Permissions page to avoid setting visible
keyboard focus for mouse users. The list does not automatically scroll
if the "Add" or "Show more" buttons are out of view. The user can use
arrow keys beyond the end of the list.

BUG: 465953
BUG: 465952


(cherry picked from commit bc8ab29f8a7a500e7aa098edb1bc13b2ee42646a)

Co-authored-by: Christoph Wolk <cwo.kde@posteo.net>