permissionstore: load unknown tables
when there are no permissions present, such as on a new user, the
initial table load won't do anything leading to us not tracking any
tables and by extension ignoring all permission changes.

instead load the table on-demand when that happens.

as a side effect we now need to issue permissionChanged on table load so
the model updates correctly

BUG: 509508