ui/Permissions: fix keyboard accessibility
When the user selects an entry from the sidebar, the Permissions page /
column receives focus, but it stops there (more precisely on the
ScrollView) and none of the controls on the page receive focus. This is
particularly bad, as the column-based approach with pages means none of
the controls are in the tab chain from elsewhere, so there is no way of
reaching them with the keyboad except the highly opaque way of
activating the page from the sidebar, then immediately pressing Backtab
from there. (Backtab at any other point will also not work, at least not
in System Settings - kcmshell might be different)

We can solve both by setting activeFocusOnTab on the page itself. This
does not focus any element by default (though this could be changed by
explicitly setting the focus property on an element), but the first
control can now be reached through simply pressing Tab, and it also
works if focus leaves the page - e.g. it can now be reached using Tab
from the search field or app list.