ui/main: improve sidebar arrow key handling
In 6.4, the sidebar (page) had different visual indicators for the entry
with keyboard focus and the element that currently has an open
permissions page, similar to how it works e.g. in Discover's sidebar.
6.5 merges the two states again, so if the user interacts with the list
using arrow keys, the highlighted item no longer matches what's
displayed in the Permissions page. In addition, using arrow down from
the search field does not affect the listview at all; if there is no
currentItem it will focus the list without selecting anything (adding
a confusing extra step where focus seems to vanish) and if an element
further down the list is the currentItem, it will immediately go there.
This is not how search fields for listviews work in most other places
(again see Discover), where pressing down always focuses the first item
in the list.

This change implements this and restores 6.4 behavior.